backend11/backend/app/requirements.txt
fastapi==0.110.3
nltk==3.8.1
numpy==1.26.4
pydantic==2.7.1
pymongo==4.7.1
sentence_transformers==2.6.1
tensorflow==2.16.1
torch==2.2.2
python-docx>=1.1.2,<1.2.0  # null
